+++ propagators/envcar/carrier_example_test.go
@@ -34,10 +34,6 @@ func ExampleCarrier_extractFromParent() {
 	fmt.Printf("Trace ID: %s\n", spanCtx.TraceID())
 	fmt.Printf("Span ID: %s\n", spanCtx.SpanID())
 	fmt.Printf("Sampled: %t\n", spanCtx.IsSampled())
-	// Output:
-	// Trace ID: 0102030405060708090a0b0c0d0e0f10
-	// Span ID: 0102030405060708
-	// Sampled: true
 }
 
 // This example is a go program where we have a trace and we'd like to inject it
@@ -77,5 +73,4 @@ func ExampleCarrier_childProcess() {
 		return
 	}
 	fmt.Print(string(out))
-	// Output: 00-0102030405060708090a0b0c0d0e0f10-0102030405060708-01
 }