+++ propagators/envcar/doc.go
@@ -2,5 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 // Package envcar implements the Environment Carrier specification as documented
-// here https://opentelemetry.io/docs/specs/otel/context/env-carriers/
+// here: https://opentelemetry.io/docs/specs/otel/context/env-carriers/.
 package envcar // import "go.opentelemetry.io/contrib/propagators/envcar"